[FIX] mail: activity view=>no tz for deadline date
In activity view, the deadline date that has no timezone is formatted
from UTC to local timezone, so if the timezone is negative, we display
by error the previous day (eg. 24 october instead of 25 october).

With this commmit, we interpret the date in local timezone so there is
no timezone issue.

note: no test added since this only happen on browser with negative
timezone which would be hard to test.

note: this was fixed similarily in 13.0 with 1d8ee9c718 but must have
been lost in a refactoring.

opw-2629681

closes odoo/odoo#78799

X-original-commit: d37b844e19b104d28bf020f7d68749f60d81923a
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>