[FIX] website: consider cover properties root element as savable
Commit [1] did adapt the selection of savable to include cover
properties detected through `[data-name="Cover"]`, but [2] did rename
the blog cover option block from `Cover` to `Blog Post Cover` therefore
it did not get detected. Because of this, only changing attributes on
the root element of the blog post cover was not marking it dirty, and
therefore not saving it unless other changes also happened (e.g. change
the picture)

Given that the name might even be translated, this commit does not use
the value of `data-name` anymore but relies on either the `s_cover` or
the `o_record_cover_container` class.

[1] https://github.com/odoo/odoo/commit/4dc6d72d4f548f7c080a79d9ac191e2dd7c5ee5f
[2] https://github.com/odoo/odoo/commit/0ec2e610ac1c1891ca9a1ef9a57560bf4d1368dd

task-2678100

Part-of: odoo/odoo#79028