[FIX] *sale*: exception when cancel SO
The commit e460195659f93c36babe289f4310e6243b463a2f fixes issue with
backorder being to quickly canceled but as it call `action_cancel`
before the condition who's looking on the SO state, the condition can
never be matched as the SO state will always be `cancel`.
So, the exception about the SO being cancelled will not be logged on
following documents.

How to reproduce:
  - Create a product and enable route MTO + Manufacture on it and create
    a BoM for this product;
  - Create a SO for this product and confirm it;
  - Cancel the SO and check the generated MO => There is no exception
    about the cancellation of the SO.

This commit fixes this issue and doesn't impact the previous fix.

task-2657799

closes odoo/odoo#79215

closes odoo/odoo#81178

Related: odoo/enterprise#22008
Related: odoo/enterprise#22814
Signed-off-by: Arnold Moyaux <arm@odoo.com>